+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedGlazedTerracottaBlock.java
@@ -1,0 +1,32 @@
+package net.geforcemods.securitycraft.blocks.reinforced;
+
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.DirectionProperty;
+import net.minecraft.world.level.material.PushReaction;
+
+public class ReinforcedGlazedTerracottaBlock extends BaseReinforcedBlock {
+	public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
+
+	public ReinforcedGlazedTerracottaBlock(Properties properties, Block vB) {
+		super(properties, vB);
+	}
+
+	@Override
+	protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+		builder.add(FACING);
+	}
+
+	@Override
+	public BlockState getStateForPlacement(BlockPlaceContext ctx) {
+		return defaultBlockState().setValue(FACING, ctx.getHorizontalDirection().getOpposite());
+	}
+
+	@Override
+	public PushReaction getPistonPushReaction(BlockState state) {
+		return PushReaction.PUSH_ONLY;
+	}
+}